Added the first sample implemention of the Netty framework. Added the ClientSocket connection code to the client
Explanation: Start the server, then start the client.
You should see that the connection is binded, connected and open.
Go to the client console and write a message, the server should receive
it and write it to console with the Channel number and IP:Port number.
It will then send the message back to the client.
You can try opening multiple client and see the creation of a unique
channel for each connection.